TensorFlow/squeezenet/src/cifar2png/requirements.txt
numpy==1.17.4
pathlib==1.0.1
Pillow==6.2.1
requests==2.22.0
six==1.13.0
tqdm==4.38.0
